+++ CHANGELOG.rst
@@ -118,6 +118,9 @@ Added
   `#308 <https://github.com/openego/eGon-data/issues/308>`_
 * Integrate supply strategies for individual and district heating
   `#232 <https://github.com/openego/eGon-data/issues/232>`_
+* Integrate industrial time series creation
+  `#237 <https://github.com/openego/eGon-data/issues/237>`_
+
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159
 
@@ -161,6 +164,8 @@ Changed
   `#281 <https://github.com/openego/eGon-data/issues/281>`_
 * Migrate mastr, mv_grid_districts and re_potential_areas to datasets
   `#297 <https://github.com/openego/eGon-data/issues/297>`_
+* Migrate industrial sites to datasets
+  `#237 <https://github.com/openego/eGon-data/issues/237>`_
 * Rename etrago tables from e.g. egon_pf_hv_bus to egon_etrago bus etc. 
   `#334 <https://github.com/openego/eGon-data/issues/334>`_
 * Move functions used by multiple datasets